test(storage): move WAL and snapshot tests to integration files
