+++ crates/storage/src/snapshot.rs
@@ -18,20 +18,3 @@ pub fn load(path: impl AsRef<Path>) -> io::Result<KvStore> {
     file.read_to_end(&mut buf)?;
     bincode::deserialize(&buf).map_err(io::Error::other)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    #[test]
-    fn snapshot_roundtrip() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join("snapshot.bin");
-        let mut store = KvStore::new();
-        store.insert(b"k".to_vec(), b"v".to_vec(), None);
-        create(&path, &store).unwrap();
-        let mut loaded = load(&path).unwrap();
-        assert_eq!(loaded.get(b"k").unwrap(), b"v".to_vec());
-    }
-}
